added props for schedule, and added preview and automate page
