+++ client/src/components/ImageQuery.js
@@ -24,6 +24,7 @@ export default function ImageQuery({image, setImage}){
           setImage(imageURL);
           setLoading(false);
         } catch (error) {
+          setLoading(false)
           console.error('Error generating image:', error);
         }
       };
@@ -35,12 +36,12 @@ export default function ImageQuery({image, setImage}){
           <textarea
             value={imageQuery}
             onChange={(e) => setImageQuery(e.target.value)}
-            placeholder="Enter prompt for LinkedIn image"
+            placeholder="Enter prompt for LinkedIn Image"
             className="w-full h-32 p-6 mt-2 bg-zinc-100 border border-zinc-300 rounded-xl resize-none font-montserrat"
           />
           <button
             onClick={handleGenerateImage}
-            className={`flex items-center justify-center px-4 py-2 mt-2 text-white rounded-full font-montserrat text-md ${loading ? 'bg-gray-400 cursor-not-allowed hover:bg-red-500' :' bg-blue-600 hover:bg-blue-700'}`}
+            className={`flex items-center justify-center px-4 py-2 mt-2 text-white rounded-full font-montserrat text-md ${loading ? 'bg-gray-400 cursor-not-allowed hover:bg-red-600' :' bg-blue-600 hover:bg-blue-700'}`}
           >
             {loading ? 'Generating Image' : 'Generate Image'}
             <FontAwesomeIcon icon={faWandMagicSparkles} className="ml-2" />
